# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Single-user local Codex-like Agent API built with Go and CloudWeGo Eino.\nExternal effects always flow through ToolRouter, EffectInference, PolicyEngine, ApprovalCenter, and Executor.",
        "title": "Local Agent API",
        "contact": {},
        "version": "dev"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/mcp/servers": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP"
                ],
                "summary": "List MCP servers",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.MCPServerListResponse"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP"
                ],
                "summary": "Create MCP server",
                "parameters": [
                    {
                        "description": "MCP server payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_tools_mcp.ServerInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_core.MCPServer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/mcp/servers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP"
                ],
                "summary": "Get MCP server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.MCPServerDetailResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP"
                ],
                "summary": "Update MCP server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "MCP server payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_tools_mcp.ServerInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_core.MCPServer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/mcp/servers/{id}/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP"
                ],
                "summary": "Refresh MCP tools",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.MCPRefreshResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/mcp/servers/{id}/test": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP"
                ],
                "summary": "Test MCP server connectivity",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.MCPTestServerResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.MCPTestServerResponse"
                        }
                    }
                }
            }
        },
        "/v1/mcp/servers/{id}/tools/{tool_name}/call": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP"
                ],
                "summary": "Call MCP tool through approval chain",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MCP server ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Tool name",
                        "name": "tool_name",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Tool call payload",
                        "name": "body",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.MCPCallToolRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ToolRouteResponse"
                        }
                    },
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.ToolRouteResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/v1/mcp/tools/{id}/policy": {
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "MCP"
                ],
                "summary": "Update MCP tool policy",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Tool policy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Policy payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_tools_mcp.ToolPolicyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/local-agent_internal_core.MCPToolPolicy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_api_handlers.LegacyErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "internal_api_handlers.LegacyErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.MCPCallToolRequest": {
            "type": "object",
            "properties": {
                "arguments": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "purpose": {
                    "type": "string"
                }
            }
        },
        "internal_api_handlers.MCPRefreshResponse": {
            "type": "object",
            "properties": {
                "state": {
                    "$ref": "#/definitions/local-agent_internal_tools_mcp.RuntimeState"
                },
                "status": {
                    "type": "string"
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/local-agent_internal_tools_mcp.MCPToolSchema"
                    }
                }
            }
        },
        "internal_api_handlers.MCPServerDetailResponse": {
            "type": "object",
            "properties": {
                "server": {
                    "$ref": "#/definitions/local-agent_internal_core.MCPServer"
                },
                "state": {
                    "$ref": "#/definitions/local-agent_internal_tools_mcp.RuntimeState"
                }
            }
        },
        "internal_api_handlers.MCPServerListResponse": {
            "type": "object",
            "properties": {
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/local-agent_internal_core.MCPServer"
                    }
                }
            }
        },
        "internal_api_handlers.MCPTestServerResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_handlers.ToolRouteResponse": {
            "type": "object",
            "properties": {
                "approval": {
                    "$ref": "#/definitions/local-agent_internal_core.ApprovalRecord"
                },
                "decision": {
                    "$ref": "#/definitions/local-agent_internal_core.PolicyDecision"
                },
                "inference": {
                    "$ref": "#/definitions/local-agent_internal_core.EffectInferenceResult"
                },
                "result": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolResult"
                }
            }
        },
...
        "local-agent_internal_core.ApprovalRecord": {
            "type": "object",
            "properties": {
                "conversation_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "decision": {
                    "$ref": "#/definitions/local-agent_internal_core.PolicyDecision"
                },
                "id": {
                    "type": "string"
                },
                "inference": {
                    "$ref": "#/definitions/local-agent_internal_core.EffectInferenceResult"
                },
                "input_snapshot": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "proposal": {
                    "$ref": "#/definitions/local-agent_internal_core.ToolProposal"
                },
                "reason": {
                    "type": "string"
                },
                "resolved_at": {
                    "type": "string"
                },
                "run_id": {
                    "type": "string"
                },
                "snapshot_hash": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/local-agent_internal_core.ApprovalStatus"
                },
                "summary": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_core.ApprovalStatus": {
            "type": "string",
            "enum": [
                "pending",
                "approved",
                "rejected"
            ],
            "x-enum-varnames": [
                "ApprovalPending",
                "ApprovalApproved",
                "ApprovalRejected"
            ]
        },
...
        "local-agent_internal_core.EffectInferenceResult": {
            "type": "object",
            "properties": {
                "approval_required": {
                    "type": "boolean"
                },
                "confidence": {
                    "type": "number"
                },
                "effects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "reason_summary": {
                    "type": "string"
                },
                "risk_level": {
                    "type": "string"
                },
                "sensitive": {
                    "type": "boolean"
                }
            }
        },
...
        "local-agent_internal_core.MCPCompatibilityProfile": {
            "type": "object",
            "properties": {
                "accept_extra_metadata": {
                    "type": "boolean"
                },
                "accept_missing_schema": {
                    "type": "boolean"
                },
                "accept_structured_result": {
                    "type": "boolean"
                },
                "accept_text_only_result": {
                    "type": "boolean"
                },
                "dialect": {
                    "type": "string"
                },
                "max_payload_bytes": {
                    "type": "integer"
                },
                "normalize_error_shape": {
                    "type": "boolean"
                },
                "strict_id_matching": {
                    "type": "boolean"
                },
                "timeout_seconds": {
                    "type": "integer"
                }
            }
        },
        "local-agent_internal_core.MCPServer": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "command": {
                    "type": "string"
                },
                "compatibility": {
                    "$ref": "#/definitions/local-agent_internal_core.MCPCompatibilityProfile"
                },
                "created_at": {
                    "type": "string"
                },
                "cwd": {
                    "type": "string"
                },
                "dialect": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "environment": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "timeout_seconds": {
                    "type": "integer"
                },
                "transport": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_core.MCPToolPolicy": {
            "type": "object",
            "properties": {
                "approval": {
                    "type": "string"
                },
                "effects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "requires_approval": {
                    "type": "boolean"
                },
                "risk_level": {
                    "type": "string"
                },
                "tool_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "local-agent_internal_core.PolicyDecision": {
            "type": "object",
            "properties": {
                "allowed": {
                    "type": "boolean"
                },
                "approval_payload": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "reason": {
                    "type": "string"
                },
                "requires_approval": {
                    "type": "boolean"
                },
                "risk_level": {
                    "type": "string"
                }
            }
        },
...
        "local-agent_internal_core.ToolProposal": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "expected_effects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "input": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "purpose": {
                    "type": "string"
                },
                "tool": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_core.ToolResult": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "output": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "started_at": {
                    "type": "string"
                },
                "tool_call_id": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_tools_mcp.CompatibilityInput": {
            "type": "object",
            "properties": {
                "accept_extra_metadata": {
                    "type": "boolean"
                },
                "accept_missing_schema": {
                    "type": "boolean"
                },
                "accept_structured_result": {
                    "type": "boolean"
                },
                "accept_text_only_result": {
                    "type": "boolean"
                },
                "dialect": {
                    "type": "string"
                },
                "max_payload_bytes": {
                    "type": "integer"
                },
                "normalize_error_shape": {
                    "type": "boolean"
                },
                "strict_id_matching": {
                    "type": "boolean"
                },
                "timeout_seconds": {
                    "type": "integer"
                }
            }
        },
        "local-agent_internal_tools_mcp.MCPToolSchema": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "input_schema": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "metadata": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_tools_mcp.RuntimeState": {
            "type": "object",
            "properties": {
                "last_refresh_at": {
                    "type": "string"
                },
                "server_id": {
                    "type": "string"
                },
                "tool_count": {
                    "type": "integer"
                },
                "tools": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/local-agent_internal_tools_mcp.MCPToolSchema"
                    }
                }
            }
        },
        "local-agent_internal_tools_mcp.ServerInput": {
            "type": "object",
            "properties": {
                "args": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "command": {
                    "type": "string"
                },
                "compatibility": {
                    "$ref": "#/definitions/local-agent_internal_tools_mcp.CompatibilityInput"
                },
                "cwd": {
                    "type": "string"
                },
                "dialect": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "environment": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "headers": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "timeout_seconds": {
                    "type": "integer"
                },
                "transport": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
        "local-agent_internal_tools_mcp.ToolPolicyInput": {
            "type": "object",
            "properties": {
                "approval": {
                    "type": "string"
                },
                "effects": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "reason": {
                    "type": "string"
                },
                "requires_approval": {
                    "type": "boolean"
                },
                "risk_level": {
                    "type": "string"
                },
                "tool_name": {
                    "type": "string"
                }
            }
        },
...
    }
}